Refactor to allow for multiple sub window plotting tabs
